+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Report cURL commands for some network-level errors (connection resets, server crashes). [#2709](https://github.com/schemathesis/schemathesis/issues/2709)
 - Cleanup pytest output. [#2714](https://github.com/schemathesis/schemathesis/issues/2714)
 - Clarify warnings and make them configurable. [#2761](https://github.com/schemathesis/schemathesis/issues/2761)
+- Rename entities related to Targeted Property-based testing. `schemathesis.target` -> `schemathesis.metric`, `TargetContext` -> `MetricContext`
 
 ### :fire: Removed